More custom op simplifications
